backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.2.0
python-magic==0.4.27
cryptography==41.0.7
phonenumbers==8.13.26
drf-spectacular==0.27.0